Make re-indexing configurable

+++ src/au/AuEncoder.h
@@ -126,6 +126,7 @@ public:
     if (clearUsageTracker) internCache_.clear();
   }
 
+  /// Removes strings that are used less than "threshold" times from the hash
   size_t purge(size_t threshold) {
     // Note: We can't modify dictInOrder_ or else the internIndex will no longer
     // match.
@@ -141,6 +142,8 @@ public:
     return purged;
   }
 
+  /// Purges the dictionary and re-idexes the remaining entries so the more
+  /// frequent ones are at the beginning (and have smaller indices).
   size_t reIndex(size_t threshold) {
     size_t purged = purge(threshold);
 
@@ -152,7 +155,8 @@ public:
 
     std::sort(dictInOrder_.begin(), dictInOrder_.end(),
               [this](const auto &a, const auto &b) {
-                return dictionary_[a].occurences < dictionary_[b].occurences;
+                // Invert comparison b/c we want frequent strings first
+                return dictionary_[a].occurences > dictionary_[b].occurences;
               });
     size_t idx = 0;
     for (auto &v : dictInOrder_) {
@@ -467,6 +471,7 @@ class Au {
   size_t records_;
   size_t purgeInterval_;
   size_t purgeThreshold_;
+  size_t reindexInterval_;
   size_t clearThreshold_;
 
   size_t tell() const {
@@ -509,8 +514,12 @@ class Au {
 
     pos_ += tracker.count();
 
-    if (records_ % purgeInterval_ == 0) {
+    if (purgeInterval_ && (records_ % purgeInterval_ == 0)) {
       purgeDictionary(purgeThreshold_);
+    }
+
+    if (reindexInterval_ && (records_ % reindexInterval_ == 0)) {
+      reIndexDictionary(purgeThreshold_);
     }
 
     if (lastDictSize_ > clearThreshold_) {
@@ -546,17 +555,22 @@ public:
 
   /**
    * @param output
-   * @param purgeInterval The dictionary will be purged after this many records
+   * @param purgeInterval The dictionary will be purged after this many records.
+   * A value of 0 means "never".
    * @param purgeThreshold Entries with a count less than this will be purged
+   * when a purge or reindex is done.
+   * @param reindexInterval The dictionary will be reindexed after this many
+   * records. A value of 0 means "never". A re-index involves a purge.
    * @param clearThreshold When the dictionary grows beyond this size, it will
    * be cleared. Large dictionaries slow down encoding.
    */
   Au(std::ostream &output, size_t purgeInterval = 250'000,
-     size_t purgeThreshold = 50, size_t clearThreshold = 1400)
+     size_t purgeThreshold = 50, size_t reindexInterval = 500'000,
+     size_t clearThreshold = 1400)
       : output_(output), outputTracker_(output),
         pos_(0), lastDictSize_(0), records_(0),
         purgeInterval_(purgeInterval), purgeThreshold_(purgeThreshold),
-        clearThreshold_(clearThreshold)
+        reindexInterval_(reindexInterval), clearThreshold_(clearThreshold)
   {
     OutputTracker outputTracker(output_);
     AuFormatter af(output_, stringIntern_);
